services/document/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
qdrant-client==1.6.9
transformers==4.35.0
tokenizers==0.14.1
numpy==1.24.3
scikit-learn==1.3.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
python-magic==0.4.27
# JWT Authentication
PyJWT==2.8.0